+++ data-frame/data-frame.go
@@ -18,16 +18,20 @@ type rowable interface {
 	String() string
 }
 
-type cell interface {
+// Cell is the interface that every cell in a DataFrame needs to comply with
+type Cell interface {
 	String() string
 	Int() (*int, error)
 	Float() (*float64, error)
 	Bool() (*bool, error)
-	NA() bool
+	NA() Cell
+	IsNA() bool
 	Checksum() [16]byte
+	Copy() Cell
 }
 
-type cells []cell
+// Cells is a wrapper for a slice of Cells
+type Cells []Cell
 
 type tointeger interface {
 	Int() (*int, error)
@@ -43,7 +47,7 @@ type tobool interface {
 
 // DataFrame is the base data structure
 type DataFrame struct {
-	columns  columns
+	Columns  columns
 	colNames []string
 	colTypes []string
 	nCols    int
@@ -53,7 +57,7 @@ type DataFrame struct {
 // C represents a way to pass Colname and Elements to a DF constructor
 type C struct {
 	Colname  string
-	Elements cells
+	Elements Cells
 }
 
 // T is used to represent the association between a column and it't type
@@ -129,7 +133,7 @@ func New(colConst ...C) (*DataFrame, error) {
 	df := &DataFrame{
 		colNames: colNames,
 		colTypes: colTypes,
-		columns:  cols,
+		Columns:  cols,
 		nRows:    colLength,
 		nCols:    len(colNames),
 	}
@@ -142,16 +146,40 @@ func (df DataFrame) Names() []string {
 	return df.colNames
 }
 
+func (df DataFrame) copy() DataFrame {
+	colnames := make([]string, len(df.colNames))
+	for k, v := range df.colNames {
+		colnames[k] = v
+	}
+	coltypes := make([]string, len(df.colTypes))
+	for k, v := range df.colTypes {
+		coltypes[k] = v
+	}
+	columns := make(columns, len(df.Columns))
+	for k, v := range df.Columns {
+		columns[k] = v.copy()
+	}
+	dfc := DataFrame{
+		colNames: colnames,
+		colTypes: coltypes,
+		Columns:  columns,
+		nRows:    df.nRows,
+		nCols:    df.nCols,
+	}
+	return dfc
+}
+
 // SetNames let us specify the column names of a DataFrame
 func (df *DataFrame) SetNames(colnames []string) error {
 	if len(df.colNames) != len(colnames) {
 		return errors.New("Different sizes for colnames array")
 	}
 
-	for k, v := range df.columns {
-		v.colName = colnames[k]
+	for k := range df.Columns {
+		df.Columns[k].colName = colnames[k]
+		df.Columns[k].recountNumChars()
+		df.colNames[k] = colnames[k]
 	}
-	df.colNames = colnames
 	return nil
 }
 
@@ -218,7 +246,7 @@ func (df *DataFrame) LoadData(records [][]string) error {
 		newDf.colTypes = append(newDf.colTypes, col.colType)
 	}
 
-	newDf.columns = cols
+	newDf.Columns = cols
 	*df = newDf
 	return nil
 }
@@ -241,41 +269,33 @@ func (df *DataFrame) LoadAndParse(records [][]string, types interface{}) error {
 		if df.nCols != len(types) {
 			return errors.New("Number of columns different from number of types")
 		}
-		for k, v := range df.columns {
-			col, err := parseColumn(v, types[k])
+		for k := range df.Columns {
+			col := df.Columns[k].copy()
+			err := col.ParseColumn(types[k])
 			if err != nil {
 				return err
 			}
 			df.colTypes[k] = col.colType
-			df.columns[k] = *col
 		}
 	case T:
 		types := types.(T)
 		for k, v := range types {
-			i, err := df.columnIndex(k)
+			i, err := df.colIndex(k)
 			if err != nil {
 				return err
 			}
-			col, err := parseColumn(df.columns[i], v)
+			col := df.Columns[*i].copy()
+			err = col.ParseColumn(v)
 			if err != nil {
 				return err
 			}
 			colIndex, _ := df.colIndex(k)
 			df.colTypes[*colIndex] = col.colType
-			df.columns[i] = *col
+			df.Columns[*colIndex] = col
 		}
 	}
 
 	return nil
-}
-
-func (df DataFrame) columnIndex(colName string) (int, error) {
-	for k, v := range df.colNames {
-		if v == colName {
-			return k, nil
-		}
-	}
-	return -1, errors.New("Index out of range")
 }
 
 // SaveRecords will save data to records in [][]string format
@@ -294,7 +314,7 @@ func (df DataFrame) SaveRecords() [][]string {
 	records = append(records, df.colNames)
 	for i := 0; i < df.nRows; i++ {
 		r := []string{}
-		for _, v := range df.columns {
+		for _, v := range df.Columns {
 			r = append(r, v.cells[i].String())
 		}
 		records = append(records, r)
@@ -314,6 +334,16 @@ func (df DataFrame) Dim() (dim [2]int) {
 	return
 }
 
+// NRows is the getter method for the number of rows in a DataFrame
+func (df DataFrame) NRows() int {
+	return df.nRows
+}
+
+// NCols is the getter method for the number of rows in a DataFrame
+func (df DataFrame) NCols() int {
+	return df.nCols
+}
+
 // colIndex tries to find the column index for a given column name
 func (df DataFrame) colIndex(colname string) (*int, error) {
 	for k, v := range df.colNames {
@@ -321,7 +351,7 @@ func (df DataFrame) colIndex(colname string) (*int, error) {
 			return &k, nil
 		}
 	}
-	return nil, errors.New("Can't find the given column")
+	return nil, errors.New("Can't find the given column:")
 }
 
 // Subset will return a DataFrame that contains only the columns and rows contained
@@ -343,12 +373,7 @@ func (df DataFrame) Subset(subsetCols interface{}, subsetRows interface{}) (*Dat
 // on the given subset
 func (df DataFrame) SubsetColumns(subset interface{}) (*DataFrame, error) {
 	// Generate a DataFrame to store the temporary values
-	newDf := DataFrame{
-		columns:  columns{},
-		nRows:    df.nRows,
-		colNames: []string{},
-		colTypes: []string{},
-	}
+	newDf := df.copy()
 
 	switch subset.(type) {
 	case R:
@@ -365,9 +390,9 @@ func (df DataFrame) SubsetColumns(subset interface{}) (*DataFrame, error) {
 		}
 
 		newDf.nCols = s.To - s.From
-		newDf.colNames = df.colNames[s.From:s.To]
-		newDf.colTypes = df.colTypes[s.From:s.To]
-		newDf.columns = df.columns[s.From:s.To]
+		newDf.colNames = newDf.colNames[s.From:s.To]
+		newDf.colTypes = newDf.colTypes[s.From:s.To]
+		newDf.Columns = newDf.Columns[s.From:s.To]
 	case []int:
 		colNums := subset.([]int)
 		if len(colNums) == 0 {
@@ -377,7 +402,7 @@ func (df DataFrame) SubsetColumns(subset interface{}) (*DataFrame, error) {
 		// Check for errors
 		colNumsMap := make(map[int]bool)
 		for _, v := range colNums {
-			if v >= df.nCols || v < 0 {
+			if v >= newDf.nCols || v < 0 {
 				return nil, errors.New("Subset out of range")
 			}
 			if _, ok := colNumsMap[v]; !ok {
@@ -387,23 +412,28 @@ func (df DataFrame) SubsetColumns(subset interface{}) (*DataFrame, error) {
 			}
 		}
 
+		cols := columns{}
+		colNames := []string{}
+		colTypes := []string{}
 		for _, v := range colNums {
-			col := df.columns[v]
-			newDf.columns = append(newDf.columns, col)
-			newDf.colNames = append(newDf.colNames, df.colNames[v])
-			newDf.colTypes = append(newDf.colTypes, df.colTypes[v])
+			col := newDf.Columns[v]
+			cols = append(cols, col)
+			colNames = append(colNames, newDf.colNames[v])
+			colTypes = append(colTypes, newDf.colTypes[v])
 		}
-		newDf.nCols = len(colNums)
+		newDf.Columns = cols
+		newDf.colNames = colNames
+		newDf.colTypes = colTypes
 	case []string:
-		columns := subset.([]string)
-		if len(columns) == 0 {
+		cols := subset.([]string)
+		if len(cols) == 0 {
 			return nil, errors.New("Empty subset")
 		}
 
-		// Check for duplicated columns
+		// Check for duplicated cols
 		colnamesMap := make(map[string]bool)
 		dupedColnames := []string{}
-		for _, v := range columns {
+		for _, v := range cols {
 			if v != "" {
 				if _, ok := colnamesMap[v]; !ok {
 					colnamesMap[v] = true
@@ -417,22 +447,23 @@ func (df DataFrame) SubsetColumns(subset interface{}) (*DataFrame, error) {
 		}
 
 		// Select the desired subset of columns
-		for _, v := range columns {
-			i, err := df.columnIndex(v)
+		columns := columns{}
+		colNames := []string{}
+		colTypes := []string{}
+		for _, v := range cols {
+			i, err := newDf.colIndex(v)
 			if err != nil {
 				return nil, err
 			}
 
-			col := df.columns[i]
-			newDf.colNames = append(newDf.colNames, v)
-			colindex, err := df.colIndex(v)
-			if err != nil {
-				return nil, err
-			}
-
-			newDf.colTypes = append(newDf.colTypes, df.colTypes[*colindex])
-			newDf.columns = append(newDf.columns, col)
+			col := newDf.Columns[*i]
+			columns = append(columns, col)
+			colNames = append(colNames, v)
+			colTypes = append(colTypes, newDf.colTypes[*i])
 		}
+		newDf.Columns = columns
+		newDf.colNames = colNames
+		newDf.colTypes = colTypes
 	default:
 		return nil, errors.New("Unknown subsetting option")
 	}
@@ -444,14 +475,8 @@ func (df DataFrame) SubsetColumns(subset interface{}) (*DataFrame, error) {
 
 // SubsetRows will return a DataFrame that contains only the selected rows
 func (df DataFrame) SubsetRows(subset interface{}) (*DataFrame, error) {
-	// TODO: Must update numChars in every column after subsetting
 	// Generate a DataFrame to store the temporary values
-	newDf := DataFrame{
-		columns:  columns{},
-		nCols:    df.nCols,
-		colNames: df.colNames,
-		colTypes: df.colTypes,
-	}
+	newDf := df.copy()
 
 	switch subset.(type) {
 	case R:
@@ -468,14 +493,16 @@ func (df DataFrame) SubsetRows(subset interface{}) (*DataFrame, error) {
 		}
 
 		newDf.nRows = s.To - s.From
-		for _, v := range df.columns {
+		columns := columns{}
+		for _, v := range newDf.Columns {
 			col, err := newCol(v.colName, v.cells[s.From:s.To])
 			if err != nil {
 				return nil, err
 			}
 			col.recountNumChars()
-			newDf.columns = append(newDf.columns, *col)
+			columns = append(columns, *col)
 		}
+		newDf.Columns = columns
 	case []int:
 		rowNums := subset.([]int)
 
@@ -485,24 +512,33 @@ func (df DataFrame) SubsetRows(subset interface{}) (*DataFrame, error) {
 
 		// Check for errors
 		for _, v := range rowNums {
-			if v >= df.nRows || v < 0 {
+			if v >= df.nRows {
 				return nil, errors.New("Subset out of range")
 			}
 		}
 
 		newDf.nRows = len(rowNums)
-		for _, v := range df.columns {
-			col, err := newCol(v.colName, nil)
+		columns := columns{}
+		for _, v := range newDf.Columns {
+			cells := Cells{}
+
+			for _, i := range rowNums {
+				if i < 0 {
+					cells = append(cells, v.empty)
+				} else {
+					cells = append(cells, v.cells[i])
+				}
+			}
+
+			col, err := newCol(v.colName, cells)
 			if err != nil {
 				return nil, err
 			}
 
-			for _, i := range rowNums {
-				col.cells = append(col.cells, v.cells[i])
-			}
 			col.recountNumChars()
-			newDf.columns = append(newDf.columns, *col)
+			columns = append(columns, *col)
 		}
+		newDf.Columns = columns
 	default:
 		return nil, errors.New("Unknown subsetting option")
 	}
@@ -511,22 +547,24 @@ func (df DataFrame) SubsetRows(subset interface{}) (*DataFrame, error) {
 }
 
 // Rbind combines the rows of two dataframes
-func Rbind(dfA DataFrame, dfB DataFrame) (*DataFrame, error) {
+func Rbind(a DataFrame, b DataFrame) (*DataFrame, error) {
+	dfa := a.copy()
+	dfb := b.copy()
 	// Check that the given DataFrame contains the same number of columns that the
 	// current dataframe.
-	if dfA.nCols != dfB.nCols {
+	if dfa.nCols != dfb.nCols {
 		return nil, errors.New("Different number of columns")
 	}
 
 	// Check that the names and the types of all columns are the same
 	colNameTypeMap := make(map[string]string)
-	for k, v := range dfA.colNames {
-		colNameTypeMap[v] = dfA.colTypes[k]
+	for k, v := range dfa.colNames {
+		colNameTypeMap[v] = dfa.colTypes[k]
 	}
 
-	for k, v := range dfB.colNames {
+	for k, v := range dfb.colNames {
 		if dfType, ok := colNameTypeMap[v]; ok {
-			if dfType != dfB.colTypes[k] {
+			if dfType != dfb.colTypes[k] {
 				return nil, errors.New("Mismatching column types")
 			}
 		} else {
@@ -535,53 +573,55 @@ func Rbind(dfA DataFrame, dfB DataFrame) (*DataFrame, error) {
 	}
 
 	cols := columns{}
-	for _, v := range dfA.colNames {
-		i, err := dfA.colIndex(v)
+	for _, v := range dfa.colNames {
+		i, err := dfa.colIndex(v)
 		if err != nil {
 			return nil, err
 		}
-		j, err := dfB.colIndex(v)
+		j, err := dfb.colIndex(v)
 		if err != nil {
 			return nil, err
 		}
-		col := dfA.columns[*i]
-		col, err = col.append(dfB.columns[*j].cells...)
+		col := dfa.Columns[*i]
+		col, err = col.append(dfb.Columns[*j].cells...)
 		if err != nil {
 			return nil, err
 		}
 		cols = append(cols, col)
 	}
-	dfA.columns = cols
-	dfA.nRows += dfB.nRows
+	dfa.Columns = cols
+	dfa.nRows += dfb.nRows
 
-	return &dfA, nil
+	return &dfa, nil
 }
 
 // Cbind combines the columns of two DataFrames
-func Cbind(dfA DataFrame, dfB DataFrame) (*DataFrame, error) {
+func Cbind(a DataFrame, b DataFrame) (*DataFrame, error) {
+	dfa := a.copy()
+	dfb := b.copy()
 	// Check that the two DataFrames contains the same number of rows
-	if dfA.nRows != dfB.nRows {
+	if dfa.nRows != dfb.nRows {
 		return nil, errors.New("Different number of rows")
 	}
 
 	// Check that the column names are unique when combined
 	colNameMap := make(map[string]bool)
-	for _, v := range dfA.colNames {
+	for _, v := range dfa.colNames {
 		colNameMap[v] = true
 	}
 
-	for _, v := range dfB.colNames {
+	for _, v := range dfb.colNames {
 		if _, ok := colNameMap[v]; ok {
 			return nil, errors.New("Conflicting column names")
 		}
 	}
 
-	dfA.colNames = append(dfA.colNames, dfB.colNames...)
-	dfA.colTypes = append(dfA.colTypes, dfB.colTypes...)
-	dfA.nCols = len(dfA.colNames)
-	dfA.columns = append(dfA.columns, dfB.columns...)
+	dfa.colNames = append(dfa.colNames, dfb.colNames...)
+	dfa.colTypes = append(dfa.colTypes, dfb.colTypes...)
+	dfa.nCols = len(dfa.colNames)
+	dfa.Columns = append(dfa.Columns, dfb.Columns...)
 
-	return &dfA, nil
+	return &dfa, nil
 }
 
 type b []byte
@@ -592,7 +632,7 @@ func uniqueRowsMap(df DataFrame) map[string]u {
 	uniqueRows := make(map[string]u)
 	for i := 0; i < df.nRows; i++ {
 		mdarr := []byte{}
-		for _, v := range df.columns {
+		for _, v := range df.Columns {
 			cs := v.cells[i].Checksum()
 			mdarr = append(mdarr, cs[:]...)
 		}
@@ -684,7 +724,7 @@ func (df DataFrame) String() (str string) {
 	if len(df.colNames) != 0 {
 		str += addLeftPadding("  ", nRowsPadding+2)
 		for k, v := range df.colNames {
-			str += addRightPadding(v, df.columns[k].numChars)
+			str += addRightPadding(v, df.Columns[k].numChars)
 			str += "  "
 		}
 		str += "\n"
@@ -692,7 +732,7 @@ func (df DataFrame) String() (str string) {
 	}
 	for i := 0; i < df.nRows; i++ {
 		str += addLeftPadding(strconv.Itoa(i)+": ", nRowsPadding+2)
-		for _, v := range df.columns {
+		for _, v := range df.Columns {
 			elem := v.cells[i]
 			str += addRightPadding(formatCell(elem), v.numChars)
 			str += "  "
@@ -714,4 +754,376 @@ func formatCell(cell interface{}) string {
 		return fmt.Sprint(val)
 	}
 	return fmt.Sprint(cell)
+}
+
+// InnerJoin returns a DataFrame containing the inner join of two other DataFrames.
+// This operation matches all rows that appear on both dataframes.
+func InnerJoin(a DataFrame, b DataFrame, keys ...string) (*DataFrame, error) {
+	dfa := a.copy()
+	dfb := b.copy()
+	// Check that we have all given keys in both DataFrames
+	errorArr := []string{}
+	for _, key := range keys {
+		ia, erra := dfa.colIndex(key)
+		ib, errb := dfb.colIndex(key)
+		if erra != nil {
+			errorArr = append(errorArr, fmt.Sprint("Can't find key \"", key, "\" on left DataFrame"))
+		}
+		if errb != nil {
+			errorArr = append(errorArr, fmt.Sprint("Can't find key \"", key, "\" on right DataFrame"))
+		}
+		// Check that the column types are the same between DataFrames
+		if ia != nil && ib != nil {
+			ta := dfa.colTypes[*ia]
+			tb := dfb.colTypes[*ib]
+			if ta != tb {
+				errorArr = append(errorArr, fmt.Sprint("Different types for key\"", key, "\". Left:", ta, " Right:", tb))
+			}
+		}
+	}
+	if len(errorArr) != 0 {
+		return nil, errors.New(strings.Join(errorArr, "\n"))
+	}
+
+	// Rename non key coumns with the same name on both DataFrames
+	colnamesa := make([]string, len(dfa.colNames))
+	colnamesb := make([]string, len(dfb.colNames))
+	for k, v := range dfa.colNames {
+		colnamesa[k] = v
+	}
+	for k, v := range dfb.colNames {
+		colnamesb[k] = v
+	}
+	for k, v := range colnamesa {
+		if idx, err := dfb.colIndex(v); err == nil {
+			if !inStringSlice(v, keys) {
+				colnamesa[k] = v + ".x"
+				colnamesb[*idx] = v + ".y"
+			}
+		}
+	}
+	dfa.SetNames(colnamesa)
+	dfb.SetNames(colnamesb)
+
+	// Get the column indexes of both columns for the given keys
+	colIdxa := []int{}
+	colIdxb := []int{}
+	for _, key := range keys {
+		ia, erra := dfa.colIndex(key)
+		ib, errb := dfb.colIndex(key)
+		if erra == nil && errb == nil {
+			colIdxa = append(colIdxa, *ia)
+			colIdxb = append(colIdxb, *ib)
+		}
+	}
+
+	// Get the combined checksum for all keys in both DataFrames
+	checksumsa := make([][]byte, dfa.nRows)
+	checksumsb := make([][]byte, dfb.nRows)
+	for _, i := range colIdxa {
+		for k, v := range dfa.Columns[i].cells {
+			b := []byte{}
+			cs := v.Checksum()
+			b = append(b, cs[:]...)
+			checksumsa[k] = append(checksumsa[k], b...)
+		}
+	}
+	for _, i := range colIdxb {
+		for k, v := range dfb.Columns[i].cells {
+			b := []byte{}
+			cs := v.Checksum()
+			b = append(b, cs[:]...)
+			checksumsb[k] = append(checksumsb[k], b...)
+		}
+	}
+
+	// Get the indexes of the rows we want to join
+	dfaIndexes := []int{}
+	dfbIndexes := []int{}
+	for ka, ca := range checksumsa {
+		for kb, cb := range checksumsb {
+			if string(ca) == string(cb) {
+				dfaIndexes = append(dfaIndexes, ka)
+				dfbIndexes = append(dfbIndexes, kb)
+			}
+		}
+	}
+
+	// Get the names of the elements that are not keys on the right DataFrame
+	nokeynamesb := []string{}
+	for _, v := range dfb.colNames {
+		if !inStringSlice(v, keys) {
+			nokeynamesb = append(nokeynamesb, v)
+		}
+	}
+
+	newdfa, _ := dfa.SubsetRows(dfaIndexes)
+	newdfb, _ := dfb.Subset(nokeynamesb, dfbIndexes)
+	return Cbind(*newdfa, *newdfb)
+}
+
+// CrossJoin returns a DataFrame containing the cartesian product of the rows on
+// both DataFrames.
+func CrossJoin(a DataFrame, b DataFrame) (*DataFrame, error) {
+	dfa := a.copy()
+	dfb := b.copy()
+	colnamesa := make([]string, len(dfa.colNames))
+	colnamesb := make([]string, len(dfb.colNames))
+	for k, v := range dfb.colNames {
+		colnamesb[k] = v
+	}
+	for k, v := range dfa.colNames {
+		if idx, err := dfb.colIndex(v); err == nil {
+			colnamesa[k] = v + ".x"
+			colnamesb[*idx] = v + ".y"
+		} else {
+			colnamesa[k] = v
+		}
+	}
+	dfa.SetNames(colnamesa)
+	dfb.SetNames(colnamesb)
+
+	// Get the indexes of the rows we want to join
+	dfaIndexes := []int{}
+	dfbIndexes := []int{}
+	for i := 0; i < dfa.nRows; i++ {
+		for j := 0; j < dfb.nRows; j++ {
+			dfaIndexes = append(dfaIndexes, i)
+			dfbIndexes = append(dfbIndexes, j)
+		}
+	}
+
+	newdfa, _ := dfa.SubsetRows(dfaIndexes)
+	newdfb, _ := dfb.SubsetRows(dfbIndexes)
+	return Cbind(*newdfa, *newdfb)
+}
+
+// LeftJoin returns a DataFrame containing the left join of two other DataFrames.
+// This operation matches all rows that appear on the left DataFrame and matches
+// it with the existing ones on the right one, filling the missing rows on the
+// right with an empty value.
+func LeftJoin(a DataFrame, b DataFrame, keys ...string) (*DataFrame, error) {
+	dfa := a.copy()
+	dfb := b.copy()
+	// Check that we have all given keys in both DataFrames
+	errorArr := []string{}
+	for _, key := range keys {
+		ia, erra := dfa.colIndex(key)
+		ib, errb := dfb.colIndex(key)
+		if erra != nil {
+			errorArr = append(errorArr, fmt.Sprint("Can't find key \"", key, "\" on left DataFrame"))
+		}
+		if errb != nil {
+			errorArr = append(errorArr, fmt.Sprint("Can't find key \"", key, "\" on right DataFrame"))
+		}
+		// Check that the column types are the same between DataFrames
+		if ia != nil && ib != nil {
+			ta := dfa.colTypes[*ia]
+			tb := dfb.colTypes[*ib]
+			if ta != tb {
+				errorArr = append(errorArr, fmt.Sprint("Different types for key\"", key, "\". Left:", ta, " Right:", tb))
+			}
+		}
+	}
+	if len(errorArr) != 0 {
+		return nil, errors.New(strings.Join(errorArr, "\n"))
+	}
+
+	// Rename non key coumns with the same name on both DataFrames
+	colnamesa := make([]string, len(dfa.colNames))
+	colnamesb := make([]string, len(dfb.colNames))
+	for k, v := range dfa.colNames {
+		colnamesa[k] = v
+	}
+	for k, v := range dfb.colNames {
+		colnamesb[k] = v
+	}
+	for k, v := range colnamesa {
+		if !inStringSlice(v, keys) {
+			if idx, err := dfb.colIndex(v); err == nil {
+				colnamesa[k] = v + ".x"
+				colnamesb[*idx] = v + ".y"
+			}
+		}
+	}
+	dfa.SetNames(colnamesa)
+	dfb.SetNames(colnamesb)
+
+	// Get the column indexes of both columns for the given keys
+	colIdxa := []int{}
+	colIdxb := []int{}
+	for _, key := range keys {
+		ia, erra := dfa.colIndex(key)
+		ib, errb := dfb.colIndex(key)
+		if erra == nil && errb == nil {
+			colIdxa = append(colIdxa, *ia)
+			colIdxb = append(colIdxb, *ib)
+		}
+	}
+
+	// Get the combined checksum for all keys in both DataFrames
+	checksumsa := make([][]byte, dfa.nRows)
+	checksumsb := make([][]byte, dfb.nRows)
+	for _, i := range colIdxa {
+		for k, v := range dfa.Columns[i].cells {
+			b := []byte{}
+			cs := v.Checksum()
+			b = append(b, cs[:]...)
+			checksumsa[k] = append(checksumsa[k], b...)
+		}
+	}
+	for _, i := range colIdxb {
+		for k, v := range dfb.Columns[i].cells {
+			b := []byte{}
+			cs := v.Checksum()
+			b = append(b, cs[:]...)
+			checksumsb[k] = append(checksumsb[k], b...)
+		}
+	}
+
+	// Get the indexes of the rows we want to join
+	dfaIndexes := []int{}
+	dfbIndexes := []int{}
+	for ka, ca := range checksumsa {
+		found := false
+		for kb, cb := range checksumsb {
+			if string(ca) == string(cb) {
+				dfaIndexes = append(dfaIndexes, ka)
+				dfbIndexes = append(dfbIndexes, kb)
+				found = true
+			}
+		}
+		if !found {
+			dfaIndexes = append(dfaIndexes, ka)
+			dfbIndexes = append(dfbIndexes, -1)
+
+		}
+	}
+
+	// Get the names of the elements that are not keys on the right DataFrame
+	nokeynamesb := []string{}
+	for _, v := range dfb.colNames {
+		if !inStringSlice(v, keys) {
+			nokeynamesb = append(nokeynamesb, v)
+		}
+	}
+
+	newdfa, _ := dfa.SubsetRows(dfaIndexes)
+	newdfb, _ := dfb.Subset(nokeynamesb, dfbIndexes)
+	return Cbind(*newdfa, *newdfb)
+}
+
+// RightJoin returns a DataFrame containing the right join of two other DataFrames.
+// This operation matches all rows that appear on the right DataFrame and matches
+// it with the existing ones on the left one, filling the missing rows on the
+// left with an empty value.
+func RightJoin(b DataFrame, a DataFrame, keys ...string) (*DataFrame, error) {
+	dfa := a.copy()
+	dfb := b.copy()
+	// Check that we have all given keys in both DataFrames
+	errorArr := []string{}
+	for _, key := range keys {
+		ia, erra := dfa.colIndex(key)
+		ib, errb := dfb.colIndex(key)
+		if erra != nil {
+			errorArr = append(errorArr, fmt.Sprint("Can't find key \"", key, "\" on left DataFrame"))
+		}
+		if errb != nil {
+			errorArr = append(errorArr, fmt.Sprint("Can't find key \"", key, "\" on right DataFrame"))
+		}
+		// Check that the column types are the same between DataFrames
+		if ia != nil && ib != nil {
+			ta := dfa.colTypes[*ia]
+			tb := dfb.colTypes[*ib]
+			if ta != tb {
+				errorArr = append(errorArr, fmt.Sprint("Different types for key\"", key, "\". Left:", ta, " Right:", tb))
+			}
+		}
+	}
+	if len(errorArr) != 0 {
+		return nil, errors.New(strings.Join(errorArr, "\n"))
+	}
+
+	// Rename non key coumns with the same name on both DataFrames
+	colnamesa := make([]string, len(dfa.colNames))
+	colnamesb := make([]string, len(dfb.colNames))
+	for k, v := range dfa.colNames {
+		colnamesa[k] = v
+	}
+	for k, v := range dfb.colNames {
+		colnamesb[k] = v
+	}
+	for k, v := range colnamesa {
+		if !inStringSlice(v, keys) {
+			if idx, err := dfb.colIndex(v); err == nil {
+				colnamesa[k] = v + ".y"
+				colnamesb[*idx] = v + ".x"
+			}
+		}
+	}
+	dfa.SetNames(colnamesa)
+	dfb.SetNames(colnamesb)
+
+	// Get the column indexes of both columns for the given keys
+	colIdxa := []int{}
+	colIdxb := []int{}
+	for _, key := range keys {
+		ia, erra := dfa.colIndex(key)
+		ib, errb := dfb.colIndex(key)
+		if erra == nil && errb == nil {
+			colIdxa = append(colIdxa, *ia)
+			colIdxb = append(colIdxb, *ib)
+		}
+	}
+
+	// Get the combined checksum for all keys in both DataFrames
+	checksumsa := make([][]byte, dfa.nRows)
+	checksumsb := make([][]byte, dfb.nRows)
+	for _, i := range colIdxa {
+		for k, v := range dfa.Columns[i].cells {
+			b := []byte{}
+			cs := v.Checksum()
+			b = append(b, cs[:]...)
+			checksumsa[k] = append(checksumsa[k], b...)
+		}
+	}
+	for _, i := range colIdxb {
+		for k, v := range dfb.Columns[i].cells {
+			b := []byte{}
+			cs := v.Checksum()
+			b = append(b, cs[:]...)
+			checksumsb[k] = append(checksumsb[k], b...)
+		}
+	}
+
+	// Get the indexes of the rows we want to join
+	dfaIndexes := []int{}
+	dfbIndexes := []int{}
+	for ka, ca := range checksumsa {
+		found := false
+		for kb, cb := range checksumsb {
+			if string(ca) == string(cb) {
+				dfaIndexes = append(dfaIndexes, ka)
+				dfbIndexes = append(dfbIndexes, kb)
+				found = true
+			}
+		}
+		if !found {
+			dfaIndexes = append(dfaIndexes, ka)
+			dfbIndexes = append(dfbIndexes, -1)
+
+		}
+	}
+
+	// Get the names of the elements that are not keys on the right DataFrame
+	nokeynamesb := []string{}
+	for _, v := range dfb.colNames {
+		if !inStringSlice(v, keys) {
+			nokeynamesb = append(nokeynamesb, v)
+		}
+	}
+
+	newdfa, _ := dfa.SubsetRows(dfaIndexes)
+	newdfb, _ := dfb.Subset(nokeynamesb, dfbIndexes)
+	return Cbind(*newdfa, *newdfb)
 }