+++ data-frame/data-frame_test.go
@@ -117,7 +117,7 @@ func TestDataFrame_LoadAndParse(t *testing.T) {
 	df := DataFrame{}
 	df.LoadAndParse(data, T{"A": "int", "C": "int"})
 	if fmt.Sprint(df.colTypes) != "[df.Int df.String df.Int df.String]" {
-		t.Error("Incorrect type parsing")
+		t.Error("Incorrect type parsing" + fmt.Sprint(df.colTypes))
 	}
 }
 
@@ -327,6 +327,51 @@ func TestDataFrame_RemoveUnique(t *testing.T) {
 	df.LoadData(data)
 
 	_, err := df.RemoveUnique()
+	if err != nil {
+		t.Error(err)
+	}
+}
+
+func TestDataFrame_Join(t *testing.T) {
+	dataa := [][]string{
+		[]string{"A", "B", "C", "D"},
+		[]string{"1", "2", "3", "4"},
+		[]string{"5", "6", "7", "8"},
+		[]string{"1", "2", "3", "4"},
+		[]string{"9", "10", "11", "12"},
+	}
+	datab := [][]string{
+		[]string{"A", "C", "F"},
+		[]string{"9", "1", "8"},
+		[]string{"9", "11", "8"},
+		[]string{"1", "3", "2"},
+		[]string{"1", "1", "2"},
+	}
+	dfa := DataFrame{}
+	dfa.LoadData(dataa)
+	dfb := DataFrame{}
+	dfb.LoadData(datab)
+	_, err := InnerJoin(dfa, dfb, "A", "X")
+	if err == nil {
+		t.Error("Should have failed: Key X not in left or right DataFrame")
+	}
+	_, err = InnerJoin(dfa, dfb, "A")
+	if err != nil {
+		t.Error(err)
+	}
+	_, err = InnerJoin(dfa, dfb, "A", "C")
+	if err != nil {
+		t.Error(err)
+	}
+	_, err = CrossJoin(dfa, dfb)
+	if err != nil {
+		t.Error(err)
+	}
+	_, err = LeftJoin(dfa, dfb, "A", "C")
+	if err != nil {
+		t.Error(err)
+	}
+	_, err = RightJoin(dfa, dfb, "C")
 	if err != nil {
 		t.Error(err)
 	}